refactor: restructure admin and dashboard pages for improved management
- Removed the old admin page and replaced it with a new dashboard structure for better event management.
- Updated the home page to redirect authenticated users to the dashboard.
- Enhanced the global dashboard to display user events and provide a more streamlined user experience.
- Implemented loading and error states for fetching user events.
- Added functionality for creating, editing, and deleting events within the admin dashboard.
