+++ frontend/pages/index.tsx
@@ -1,13 +1,10 @@
 import React from "react";
 import { useEffect, useState, useContext } from 'react';
 import Link from 'next/link';
+import { useRouter } from 'next/router';
 import { UserContext } from './_app';
 import { Card } from "../components/ui/card";
 import { UserRegistrationForm } from '../components/UserRegistrationForm';
-import { QuickStats } from "../components/shared/QuickStats";
-import { EventCard } from "../components/shared/EventCard";
-import { ActivityFeed } from "../components/shared/ActivityFeed";
-import { JoinEventWidget } from "../components/shared/JoinEventWidget";
 
 // Define the user type based on how it's used in the component
 interface User {
@@ -29,6 +26,7 @@ interface Event {
 }
 
 export default function Home() {
+  const router = useRouter();
   const [firstUserNeeded, setFirstUserNeeded] = useState(false);
   const [firstUserError, setFirstUserError] = useState('');
   const [firstUserSuccess, setFirstUserSuccess] = useState('');
@@ -40,9 +38,7 @@ export default function Home() {
   const [eventsLoading, setEventsLoading] = useState(false);
   const [eventsError, setEventsError] = useState('');
 
-  const [userEvents, setUserEvents] = useState<Event[] | null>(null);
-  const [userEventsLoading, setUserEventsLoading] = useState(false);
-  const [userEventsError, setUserEventsError] = useState('');
+
 
   useEffect(() => {
     fetch((process.env.NEXT_PUBLIC_API_URL || 'http://localhost:4000') + '/', {
@@ -83,94 +79,14 @@ export default function Home() {
     fetchSettingsAndEvents();
   }, []);
 
-  // Authenticated user: fetch their events
+  // Redirect authenticated users to dashboard
   useEffect(() => {
     if (user && !firstUserNeeded) {
-      setUserEventsLoading(true);
-      fetch((process.env.NEXT_PUBLIC_API_URL || 'http://localhost:4000') + '/api/users/me/events', {
-        credentials: 'include',
-      })
-        .then(res => res.json())
-        .then(data => {
-          setUserEvents(data.events || []);
-          setUserEventsLoading(false);
-        })
-        .catch(() => {
-          setUserEventsError('Could not load your events.');
-          setUserEventsLoading(false);
-        });
+      router.push('/dashboard');
     }
-  }, [user, firstUserNeeded]);
+  }, [user, firstUserNeeded, router]);
 
-  // Show empty state if user has no events
-  if (user && userEvents && userEvents.length === 0) {
-    return (
-      <div className="min-h-screen flex flex-col items-center justify-center bg-background transition-colors duration-200 p-4">
-        <Card className="w-full max-w-md text-center p-4 sm:p-8">
-          <h1 className="text-2xl font-bold mb-4 text-foreground">No Events Yet</h1>
-          <p className="mb-4 text-muted-foreground">You have not been added to any events yet. Please check your email for invites or contact an event organizer.</p>
-          <Link href="/profile/events" className="text-primary hover:underline font-semibold">Manage Event Invites</Link>
-        </Card>
-      </div>
-    );
-  }
 
-  // Show loading state for user events
-  if (user && userEventsLoading) {
-    return (
-      <div className="min-h-screen flex flex-col items-center justify-center bg-background transition-colors duration-200 p-4">
-        <Card className="w-full max-w-md text-center p-4 sm:p-8">
-          <p className="text-muted-foreground">Loading your events...</p>
-        </Card>
-      </div>
-    );
-  }
-
-  // Show error state for user events
-  if (user && userEventsError) {
-    return (
-      <div className="min-h-screen flex flex-col items-center justify-center bg-background transition-colors duration-200 p-4">
-        <Card className="w-full max-w-md text-center p-4 sm:p-8">
-          <p className="text-destructive">{userEventsError}</p>
-        </Card>
-      </div>
-    );
-  }
-
-  // Show global dashboard for user with one or more events
-  if (user && userEvents && userEvents.length > 0) {
-    return (
-      <div className="min-h-screen flex flex-col items-center justify-center bg-background transition-colors duration-200 p-4">
-        <QuickStats />
-        <ActivityFeed />
-        <JoinEventWidget onJoin={() => {
-          // Refetch user events after joining
-          setUserEventsLoading(true);
-          fetch((process.env.NEXT_PUBLIC_API_URL || 'http://localhost:4000') + '/api/users/me/events', {
-            credentials: 'include',
-          })
-            .then(res => res.json())
-            .then(data => {
-              setUserEvents(data.events || []);
-              setUserEventsLoading(false);
-            })
-            .catch(() => {
-              setUserEventsError('Could not load your events.');
-              setUserEventsLoading(false);
-            });
-        }} />
-        <Card className="w-full max-w-md text-center p-4 sm:p-8 mb-6">
-          <h1 className="text-2xl font-bold mb-4">Your Global Dashboard</h1>
-          <p className="mb-4 text-muted-foreground">Here are your events. (Full dashboard UI coming next!)</p>
-        </Card>
-        <div className="w-full max-w-md flex flex-col gap-4">
-          {userEvents.map(event => (
-            <EventCard key={event.id} event={event} />
-          ))}
-        </div>
-      </div>
-    );
-  }
 
   if (firstUserNeeded) {
     return (
